Move `Name` out of `bevy_core`
