+++ crates/bevy_hierarchy/src/valid_parent_check_plugin.rs
@@ -53,7 +53,7 @@ impl<T> Default for ReportHierarchyIssue<T> {
 /// (See B0004 explanation linked in warning message)
 pub fn check_hierarchy_component_has_valid_parent<T: Component>(
     parent_query: Query<
-        (Entity, &Parent, Option<&bevy_core::Name>),
+        (Entity, &Parent, Option<&Name>),
         (With<T>, Or<(Changed<Parent>, Added<T>)>),
     >,
     component_query: Query<(), With<T>>,
